+++ src/main.cpp
@@ -15,6 +15,7 @@ extern const luaL_Reg battle_functions[];
 extern const luaL_Reg campaign_ui_functions[];
 extern const luaL_Reg twdll_core[];
 
+extern void register_character_methods(lua_State *L);
 extern void register_faction_methods(lua_State *L);
 extern void register_unit_methods(lua_State *L);
 extern void register_military_force_methods(lua_State *L);
@@ -76,6 +77,10 @@ extern "C" __declspec(dllexport) int luaopen_twdll(lua_State *L) {
 
     l_register(L, "twdll", twdll_core);
     l_setfield(L, -2, "core");
+
+    l_register(L, "twdll_character", character_functions);
+    l_setfield(L, -2, "character");
+    register_character_methods(L);
 
     l_register(L, "twdll_unit", unit_functions);
     l_setfield(L, -2, "unit");